+++ 1612001_1612009/src/eptipi.h
@@ -63,7 +63,7 @@ protected:
 		std::string path = "";
 		Eptipi * mainFTP = NULL;
 		CSocket * dataCon = NULL;
-		int filesize = 0;
+		INT64 filesize = 0;
 	};
 
 	CSocket * openActivePortAndConnect();
